Fix menu icon clipping: use int16_t coordinates in drawXBMP (#13)
* Initial plan

* Fix menu icon clipping by changing drawXBMP coordinates to int16_t


---------

+++ src/NullDisplayManager.h
@@ -103,7 +103,7 @@ public:
         // Silent no-op
     }
 
-    void drawXBMP(uint8_t x, uint8_t y, uint8_t w, uint8_t h, const uint8_t* bitmap) override {
+    void drawXBMP(int16_t x, int16_t y, uint8_t w, uint8_t h, const uint8_t* bitmap) override {
         // Silent no-op
     }
